Fix appearance of focus ring around header logo
The original styling of `hugo-theme-terminal` doesn't include any margin
between the rectangular site logo and the repeating vertical stripes in
the main header. This causes the logo to appear slightly asymmetrical on
the right side vs. the left side, which ruins the appearance of the
focus ring around the logo when the user is `Tab`-ing around the site.

We apply an 8px left margin to `.header-logo-container::after`, which
successfully fixes this issue. It also makes the distance between the
orange stripes look more even on mobile screens, which is a nice bonus.