Trim overflowing codeblock "name=" text with ellipsis
Because we're working with pseudo-elements using a combo of absolute and
relative positioning, instead of proper HTML elements with flexbox[^1],
we have to make some ugly approximations in our CSS to make this happen.
At the very least, the results should be "good enough" for most mobile
devices and browsers I've tested so far.

[^1]: This is an unfortunate a limitation of Zola, which generates very
      minimal HTML for Markdown code blocks.