Fix responsive_image shortcode producing blurry images
If the `responsive_image()` shortcode is passed a source image that is
already low-resolution from the start (specifically, larger than the
max-width of the "center" and "left" site layouts - 768px wide - but not
_quite_ large enough to satisfy the next width available - 1280px wide),
we accidentally end up capping out with a 640px wide image when we
could've also included a 768px one instead. This is fixed adding a 768
width option to `config.extra.responsive_images.widths = []`.

Furthermore, the media queries in our `sizes=` attribute string are
somewhat wrong for the "full-width" site layout, so we loosen the
constraints a little if that setting is enabled.

Finally, we bump up the image quality of resized lossy images to 80%
(seems to keep our 2560-width highest resolution test image at a decent
~1.1 MB, while keeping everything else firmly in the kilobytes range) to
improve the quality of JPEGs that are already low resolution.