Set explicit width and height for responsive images
This commit improves the loading experience for responsive images (and
videos) by setting an initial width and height for these assets so the
browser can infer the image's aspect ratio before lazy-loading. This, in
turn, prevents the content from reflowing as the image "pops in".

This technique comes from this article, which I found via a Google
Lighthouse analysis of the page:

https://web.dev/articles/optimize-cls#images_without_dimensions