Improve responsiveness of mobile menu
This small script improves the responsive behavior of the hamburger menu
slightly by automatically calling `menu.hidePopover()` if the user
enlarges the browser window while the menu is still open. I wish there
was a way to specify this in CSS, but apparently not...

Curiously enough, it seems I'm not the first person to run into this
issue. Someone else on r/webdev had asked 8 months ago whether this
"hack" is considered good practice, but alas received no reply.

https://www.reddit.com/r/webdev/comments/1gluxzu/html_popover_api_menu_slot_into_page_layout_on/