Add rel="tag" attribute to tag taxonomy links
This attribute is a microformat which indicates that a hyperlink refers
to a "tag" describing the current HTML document. See documentation below:

https://html.spec.whatwg.org/multipage/links.html#link-type-tag
https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Attributes/rel#tag
https://microformats.org/wiki/rel-tag

Note that `rel="tag"` links must describe _the entire page_, not just a
portion, so we can't use them in Zola "sections" or "taxonomy" pages,
only in full-length blog post pages.