Display value of config.title in header by default
This theme will now default to displaying the official site title (the
contents of `config.title`) in the main header, without needing any
input from the user. This behavior can be overridden by the user via the
`config.extra.header_title` key which provides an alternative string to
display instead.